+++ src/irvine/oeis/a003/A003830.java
@@ -4,16 +4,16 @@ import irvine.math.z.Z;
 import irvine.oeis.Sequence;
 
 /**
- * A003830 Order of universal Chevalley group <code>D_n (3)</code>.
+ * A003830 Order of universal Chevalley group D_n (3).
  * @author Sean A. Irvine
  */
 public class A003830 implements Sequence {
 
-  private int mN = 0;
+  private long mN = 0;
 
-  static Z orderUniversalChevalleyD(final Z q, final int n) {
+  static Z orderUniversalChevalleyD(final Z q, final long n) {
     Z p = q.pow(n * (n - 1)).multiply(q.pow(n).subtract(1));
-    for (int k = 1; k < n; ++k) {
+    for (long k = 1; k < n; ++k) {
       p = p.multiply(q.pow(2 * k).subtract(1));
     }
     return p;
